docs: Add api-examples for `export_test_values()` and `snapshot_preprocess_input()`
